+++ skbio/tree/_tree.py
@@ -2716,6 +2716,71 @@ class TreeNode(SkbioObject):
 
         return dist_f(self_matrix, other_matrix)
 
+    @experimental(as_of="0.4.1-dev")
+    def bifurcate(self, insert_length=None):
+        r"""Reorders the tree into a bifurcating tree.
+
+        All nodes that have more than 2 children will
+        have additional intermediate nodes inserted to ensure that
+        every node has only 2 children.
+
+        Parameters
+        ----------
+        insert_length : int, optional
+            The branch length assigned to all inserted nodes.
+
+        See Also
+        --------
+        prune
+
+        Notes
+        -----
+        Any nodes that have a single child can be collapsed using the
+        prune method to create strictly bifurcating trees.
+
+        Examples
+        --------
+        >>> from skbio import TreeNode
+        >>> tree = TreeNode.read(["((a,b,g,h)c,(d,e)f)root;"])
+        >>> print(tree.ascii_art())
+                            /-a
+                           |
+                           |--b
+                  /c-------|
+                 |         |--g
+                 |         |
+        -root----|          \-h
+                 |
+                 |          /-d
+                  \f-------|
+                            \-e
+        >>> tree.bifurcate()
+        >>> print(tree.ascii_art())
+                            /-h
+                  /c-------|
+                 |         |          /-g
+                 |          \--------|
+                 |                   |          /-a
+        -root----|                    \--------|
+                 |                              \-b
+                 |
+                 |          /-d
+                  \f-------|
+                            \-e
+        """
+        for n in self.traverse(include_self=True):
+            if len(n.children) > 2:
+                stack = n.children
+                while len(stack) > 2:
+                    ind = stack.pop()
+                    intermediate = TreeNode()
+                    intermediate.length = insert_length
+                    intermediate.extend(stack)
+                    n.append(intermediate)
+                    for k in stack:
+                        n.remove(k)
+                    n.extend([ind, intermediate])
+
     @experimental(as_of="0.4.0")
     def index_tree(self):
         """Index a tree for rapid lookups within a tree array